refactor(interp): move tier-up and retirement policy out of the compiler
jit.go is the architecture-neutral JIT driver, but it also declared the
interpreter's tier-up and retirement policy: counters, watchdog, newWatchdog,
givesUp, nativeFrameLimit, loopBudget, retireWindow, retireGiveUpThreshold.

None of that is compilation. Retirement rewrites i.code, i.natives, and i.cold
- it is dispatch policy, owned by the interpreter. docs/jit-internals.md
already said so ("Retirement is runtime-side... only ever mutates the local
interpreter's dispatch table") while the code sat in the compiler's file.

The backend never reads any of them; their consumers are all in interp.go. So
keeping them in the compiler would have forced eight symbols to be exported
when the compiler moves into its own package, for nothing. Measuring that move
is what surfaced this: of the 98 symbols a naive relocation would have had to
export, thirteen were never the compiler's to begin with.

They now live in interp/tier.go. Pure relocation, no logic touched.

Also correct two doc claims: jit-internals.md's source-of-truth table gains the
new owner, and testing.md's coverage-gap entry pointed at jit.go for
checkRetire, which has always been in interp.go.
